# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ghana School Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /classes:
    get:
      operationId: listClasses
      tags: [classes]
      summary: List classes
      responses:
        "200":
          description: List of classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SchoolClass"
    post:
      operationId: createClass
      tags: [classes]
      summary: Create class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassBody"
      responses:
        "201":
          description: Created class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchoolClass"
  /classes/{id}:
    put:
      operationId: updateClass
      tags: [classes]
      summary: Update class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClassBody"
      responses:
        "200":
          description: Updated class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchoolClass"
    delete:
      operationId: deleteClass
      tags: [classes]
      summary: Delete class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /subjects:
    get:
      operationId: listSubjects
      tags: [subjects]
      summary: List subjects
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Create subject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubjectBody"
      responses:
        "201":
          description: Created subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /subjects/{id}:
    put:
      operationId: updateSubject
      tags: [subjects]
      summary: Update subject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubjectBody"
      responses:
        "200":
          description: Updated subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    delete:
      operationId: deleteSubject
      tags: [subjects]
      summary: Delete subject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List students
      parameters:
        - name: classId
          in: query
          required: false
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentBody"
      responses:
        "201":
          description: Created student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
    put:
      operationId: updateStudent
      tags: [students]
      summary: Update student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentBody"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /teachers:
    get:
      operationId: listTeachers
      tags: [teachers]
      summary: List teachers
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of teachers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Teacher"
    post:
      operationId: createTeacher
      tags: [teachers]
      summary: Create teacher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeacherBody"
      responses:
        "201":
          description: Created teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
  /teachers/{id}:
    get:
      operationId: getTeacher
      tags: [teachers]
      summary: Get teacher
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
        "404":
          description: Not found
    put:
      operationId: updateTeacher
      tags: [teachers]
      summary: Update teacher
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeacherBody"
      responses:
        "200":
          description: Updated teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
    delete:
      operationId: deleteTeacher
      tags: [teachers]
      summary: Delete teacher
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /class-assignments:
    get:
      operationId: listClassAssignments
      tags: [classes]
      summary: List class-teacher and class-subject assignments
      parameters:
        - name: classId
          in: query
          required: false
          schema:
            type: integer
        - name: academicYearId
          in: query
          required: false
          schema:
            type: integer
        - name: termId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of class assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClassAssignment"
    post:
      operationId: createClassAssignment
      tags: [classes]
      summary: Assign teacher to class/subject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassAssignmentBody"
      responses:
        "201":
          description: Created assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassAssignment"
  /class-assignments/{id}:
    delete:
      operationId: deleteClassAssignment
      tags: [classes]
      summary: Remove class assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    SchoolClass:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        level:
          type: string
          enum: [nursery, kg1, kg2, primary1, primary2, primary3, primary4, primary5, primary6, jhs1, jhs2, jhs3]
        section:
          type: ["string", "null"]
        capacity:
          type: ["integer", "null"]
        studentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, level, studentCount, createdAt]
    CreateClassBody:
      type: object
      properties:
        name:
          type: string
        level:
          type: string
          enum: [nursery, kg1, kg2, primary1, primary2, primary3, primary4, primary5, primary6, jhs1, jhs2, jhs3]
        section:
          type: string
        capacity:
          type: integer
      required: [name, level]
    UpdateClassBody:
      type: object
      properties:
        name:
          type: string
        level:
          type: string
          enum: [nursery, kg1, kg2, primary1, primary2, primary3, primary4, primary5, primary6, jhs1, jhs2, jhs3]
        section:
          type: string
        capacity:
          type: integer
    Subject:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, name, createdAt]
    CreateSubjectBody:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        description:
          type: string
      required: [name]
    UpdateSubjectBody:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        description:
          type: string
    Student:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        otherNames:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
          format: date
        gender:
          type: string
          enum: [male, female]
        classId:
          type: ["integer", "null"]
        className:
          type: ["string", "null"]
        admissionDate:
          type: string
          format: date
        status:
          type: string
          enum: [active, inactive, graduated, transferred]
        parentName:
          type: ["string", "null"]
        parentPhone:
          type: ["string", "null"]
        parentEmail:
          type: ["string", "null"]
        parentRelationship:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, studentId, firstName, lastName, gender, admissionDate, status, createdAt]
    CreateStudentBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        otherNames:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female]
        classId:
          type: integer
        admissionDate:
          type: string
          format: date
        status:
          type: string
          enum: [active, inactive, graduated, transferred]
        parentName:
          type: string
        parentPhone:
          type: string
        parentEmail:
          type: string
        parentRelationship:
          type: string
        address:
          type: string
      required: [firstName, lastName, gender, admissionDate]
    UpdateStudentBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        otherNames:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female]
        classId:
          type: integer
        admissionDate:
          type: string
          format: date
        status:
          type: string
          enum: [active, inactive, graduated, transferred]
        parentName:
          type: string
        parentPhone:
          type: string
        parentEmail:
          type: string
        parentRelationship:
          type: string
        address:
          type: string
    Teacher:
      type: object
      properties:
        id:
          type: integer
        staffId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        qualification:
          type: ["string", "null"]
        specialization:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        hireDate:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
      required: [id, staffId, firstName, lastName, status, createdAt]
    CreateTeacherBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        email:
          type: string
        qualification:
          type: string
        specialization:
          type: string
        status:
          type: string
          enum: [active, inactive]
        hireDate:
          type: string
          format: date
      required: [firstName, lastName]
    UpdateTeacherBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        email:
          type: string
        qualification:
          type: string
        specialization:
          type: string
        status:
          type: string
          enum: [active, inactive]
        hireDate:
          type: string
          format: date
    ClassAssignment:
      type: object
      properties:
        id:
          type: integer
        classId:
          type: integer
        className:
          type: ["string", "null"]
        teacherId:
          type: integer
        teacherName:
          type: ["string", "null"]
        subjectId:
          type: ["integer", "null"]
        subjectName:
          type: ["string", "null"]
        academicYearId:
          type: integer
        termId:
          type: ["integer", "null"]
        isClassTeacher:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, classId, teacherId, academicYearId, isClassTeacher, createdAt]
    CreateClassAssignmentBody:
      type: object
      properties:
        classId:
          type: integer
        teacherId:
          type: integer
        subjectId:
          type: integer
        academicYearId:
          type: integer
        termId:
          type: integer
        isClassTeacher:
          type: boolean
      required: [classId, teacherId, academicYearId]
...